only suhu have type double in diagnosa form

+++ core/src/main/java/com/example/core/data/remote/network/ApiService.kt
@@ -21,9 +21,9 @@ interface ApiService {
         @Field("alamat") address: String,
         @Field("usia") age: Int,
         @Field("tingkat_kesadaran") kesadaran: String,
-        @Field("pernafasan") pernafasan: Double,
-        @Field("denyut_nadi") denyutNadi: Double,
-        @Field("tekanan_darah") tekananDarah: Double,
+        @Field("pernafasan") pernafasan: Int,
+        @Field("denyut_nadi") denyutNadi: Int,
+        @Field("tekanan_darah") tekananDarah: Int,
         @Field("suhu") suhu: Double,
     ): DiagnosaResponse
